KP-7713 Add CSS class for LBR message templates
Previously we have used wordpress-builtins
`has-pale-pink-backgroud-color` and `has-background` for otherwise
normal `p` elements. This causes problems though: multi-paragraph
messages each have their own background box. Even if those classes were
applied to a container `div` or such, the spacing becomes wonky, as the
margins and paddings defined for `p` and `has-backgorund` elements
together cause a really big empty area at the bottom of the message
template box.

This is fixed by creating a separate class for this specific use. The
background colour is kept as is, but the empty space around the content
is now created by the container element, meaning that the `p`-elements
within don't need to have classes of their own. This will hopefully also
make the page more likely to not break when edited with the visual
editor.

The double margin/padding is prevented by removing the bottom margin
from the last `p` in the message container.