KP-9064 Remove unused modernizr
There does not seem to be any use for modernizer apart from including it
in header.php: e.g. `grep -r oderniz` in /var/www/html of preprod-proxy
yields no relevant results (only the modernizer file itself and some
mentions to it in comments of packages inspired by it), so it _seems_
safe to delete. Quick testing of the portal pages without it available
did not reveal any problems either.