KP-9145 Fix unreachable machine after attempting executing firewall
The currently available Pouta image is AlmaLinux 9.5, while the earlier
production and preprod machines run on 9.4. In addition to this
resulting in iptables no longer being available (fixed in fd33e51),
this also means that /sbin/service is not there when our playbook tries
to execute the firewall.

From the Ansible user point of view this results just in the playbook
execution freezing (until a connection timeout much later) and the
machien being unreachable via SSH etc, but attempting to run the
relevant command manually from the command line yields a more
informative error message:
```
[almalinux@portal-pre-prod ~]$ sudo /etc/rc.d/rc.firewall restart
/etc/rc.d/rc.firewall: line 6: /sbin/service: No such file or directory
```

To fix this, we either need to use an alternative method to start the
service or explicitly install the package that provides `/sbin/service`.
The latter approach was chosen here to keep the changes to a minimum.