KP-9145 Stop relying on python -> python3 symlink
Default `/usr/bin/python` -> `/usr/bin/python3` symlink is no longer
available in the new AlmaLinux 9.5 images on Pouta, which caused our pid
generator to crash on startup. The actual Python version has stayed the
same though:
```
[almalinux@portal-PROD ~]$ python --version
Python 3.9.21
```
```
[almalinux@portal-pre-prod ~]$ python --version
Python 3.9.21
```

We could create the symlink ourselves, but it seemed better to follow
the explicit versioning chosen for the rest of the system, so the
shebangs were updated to point to the executable directly.