KP-9145 Add AFFIL=FIN-CLARIN laundry tag to license template (#22)
* KP-9145 Fix whitespace & indentation to 4 spaces

This matches https://www.php-fig.org/psr/psr-2/ requirement
> Code MUST use 4 spaces for indenting, not tabs.

Also fixed some other minor style stuff, like opening curly braces never
being on their own line.

* KP-9145 Fix default condition merging bug

Earlier it could e.g. happen that you tick boxes ID and AFFIL=EDU for a
RES resource, but the license page would show ID and PLAN. Similarly, if
you ticked ID, AFFIL=EDU and PLAN, you would get ID, PLAN, PLAN.

This was due to use of the `+` operator for arrays, which is an "union",
but PHP arrays are more like dicts and the union works with the keys of
the array, which in our case are implicit indexing. This example and its
output illustrate the behaviour:
```
$a=array('a', 'b');
$b=array('c', 'd');

print_r($a);
print_r($b);
print_r($a + $b);
```
```
Array
(
    [0] => a
    [1] => b
)
Array
(
    [0] => c
    [1] => d
)
Array
(
    [0] => a
    [1] => b
)
```

Now the behaviour has been fixed so that we do union of the values in
the array, resulting in the desired behaviour.

NB: this implementation means that the order of the displayed elements
can change. I think this also happened with the old implementation
though. If you for example:
1. Create a RES license with ID and PLAN selected in the identification
   and access conditions, save page and view it, you see them in order
   ID, PLAN.
2. Add AFFIL=EDU, save and f5, you see ID, PLAN, AFFIL=EDU
3. Add FF and save, you now see ID, AFFIL=EDU, FF, PLAN.
4. Remove FF and save, you now see ID, AFFIL=EDU, PLAN (effectively same
   license as in step 2, but different order).

* KP-9145 Add AFFIL=FIN-CLARIN license text

* KP-9145 Switch to URNs, add en/fi language variants

---------

Co-authored-by: Martin Matthiesen <martin.matthiesen@csc.fi>