KP-8293 Do not overwrite existing backup files when restoring
The old implementation copied the backup file to be restored on top of
the today's backup file. This could lead to loss of important data e.g.
if local backup file is from an earlier week or the IP addresses in the
inventory are incorrect (overwriting production data with preprod data).

Now the backup is copied into a temporary file instead and restored from
there (this requires changes from
https://github.com/CSCfi/Kielipankki-ansible-common/pull/30), so there's
no risk of losing the original backup files.